ddb/db_print.c: use unsigned long instead of an unsigned int
Variable db_maxoff is in other files declared as unsigned long.

* ddb/db_print.c (db_maxoff): Use unsigned long instead of an unsigned int.

+++ ddb/db_print.c
@@ -53,7 +53,7 @@
 #include <ddb/db_task_thread.h>
 #include <ddb/db_print.h>
 
-extern unsigned int	db_maxoff;
+extern unsigned long	db_maxoff;
 
 /* ARGSUSED */
 void
